deployment/deploy.py
"""
Deployment Script for Vertex AI Agent Engine

Deploys the document extraction engine to Vertex AI Agent Engine.
"""

import os
import sys
import logging
import argparse
import yaml
from pathlib import Path
from typing import Dict, Any, Optional

# Add parent directory to path for imports
sys.path.insert(0, str(Path(__file__).parent.parent))

logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)


class AgentEngineDeployer:
    """
    Handles deployment to Vertex AI Agent Engine
    """

    def __init__(
        self,
        project_id: str,
        region: str = "us-central1",
        config_path: str = "config/agent_config.yaml"
    ):
        self.project_id = project_id
        self.region = region
        self.config_path = config_path

        # TODO: Import actual Google Cloud libraries
        # from google.cloud import aiplatform
        # from google.cloud.aiplatform import Agent

        logger.info(f"Deployer initialized for project: {project_id}")

    def load_config(self) -> Dict[str, Any]:
        """
        Load agent configuration from YAML

        Returns:
            Configuration dictionary
        """
        logger.info(f"Loading configuration from: {self.config_path}")

        with open(self.config_path, 'r') as f:
            config = yaml.safe_load(f)

        # Substitute environment variables
        config = self._substitute_env_vars(config)

        return config

    def _substitute_env_vars(self, config: Any) -> Any:
        """
        Recursively substitute ${VAR} with environment variables

        Args:
            config: Configuration object (dict, list, or str)

        Returns:
            Config with substituted values
        """
        if isinstance(config, dict):
            return {k: self._substitute_env_vars(v) for k, v in config.items()}
        elif isinstance(config, list):
            return [self._substitute_env_vars(item) for item in config]
        elif isinstance(config, str):
            # Replace ${VAR} patterns
            import re
            def replace_var(match):
                var_name = match.group(1)
                return os.environ.get(var_name, match.group(0))

            return re.sub(r'\$\{([^}]+)\}', replace_var, config)
        else:
            return config

    def validate_prerequisites(self) -> bool:
        """
        Validate that all prerequisites are met

        Returns:
            True if all prerequisites are met
        """
        logger.info("Validating prerequisites...")

        checks = []

        # Check Google Cloud SDK
        # checks.append(self._check_gcloud_installed())

        # Check credentials
        # checks.append(self._check_credentials())

        # Check project exists
        # checks.append(self._check_project_exists())

        # Check APIs enabled
        # checks.append(self._check_apis_enabled())

        # Placeholder
        checks.append(True)

        if all(checks):
            logger.info("✅ All prerequisites met")
            return True
        else:
            logger.error("❌ Some prerequisites failed")
            return False

    def upload_schemas(self, bucket_name: str) -> None:
        """
        Upload schema files to GCS

        Args:
            bucket_name: GCS bucket name
        """
        logger.info(f"Uploading schemas to gs://{bucket_name}/schemas/")

        # TODO: Implement GCS upload
        # from google.cloud import storage
        # client = storage.Client()
        # bucket = client.bucket(bucket_name)
        #
        # schema_dir = Path("schemas")
        # for schema_file in schema_dir.glob("*.json"):
        #     blob = bucket.blob(f"schemas/{schema_file.name}")
        #     blob.upload_from_filename(schema_file)
        #     logger.info(f"Uploaded {schema_file.name}")

        logger.info("Schema upload completed")

    def deploy_agent(self) -> Dict[str, Any]:
        """
        Deploy agent to Agent Engine

        Returns:
            Deployment info including endpoint URL
        """
        logger.info("🚀 Deploying to Vertex AI Agent Engine...")

        config = self.load_config()

        # TODO: Implement actual deployment
        # from google.cloud import aiplatform
        # from google.cloud.aiplatform import Agent
        #
        # aiplatform.init(project=self.project_id, location=self.region)
        #
        # agent = Agent.create(
        #     display_name=config['metadata']['name'],
        #     agent_config=config['spec'],
        #     description="Multi-agent document extraction engine for financial documents"
        # )
        #
        # logger.info(f"✅ Agent created: {agent.resource_name}")
        #
        # # Create serving endpoint
        # endpoint = agent.deploy(
        #     deployed_model_display_name=f"{config['metadata']['name']}-v1",
        #     traffic_percentage=100,
        #     sync=True
        # )
        #
        # logger.info(f"🌐 Endpoint deployed: {endpoint.resource_name}")

        # Placeholder return
        return {
            'agent_name': config['metadata']['name'],
            'project_id': self.project_id,
            'region': self.region,
            'endpoint_url': f"https://{self.region}-aiplatform.googleapis.com/v1/projects/{self.project_id}/locations/{self.region}/endpoints/placeholder",
            'status': 'deployed'
        }

    def create_endpoint(self, agent_resource_name: str) -> str:
        """
        Create serving endpoint for agent

        Args:
            agent_resource_name: Resource name of deployed agent

        Returns:
            Endpoint resource name
        """
        logger.info("Creating serving endpoint...")

        # TODO: Implement endpoint creation

        return f"projects/{self.project_id}/locations/{self.region}/endpoints/placeholder"

    def test_deployment(self, endpoint_url: str) -> bool:
        """
        Test deployed endpoint with sample request

        Args:
            endpoint_url: Endpoint URL to test

        Returns:
            True if test successful
        """
        logger.info("Testing deployment...")

        # TODO: Implement test request

        logger.info("✅ Deployment test passed")
        return True

    def deploy(self, bucket_name: Optional[str] = None) -> Dict[str, Any]:
        """
        Full deployment workflow

        Args:
            bucket_name: GCS bucket for schemas (optional)

        Returns:
            Deployment information
        """
        logger.info("=" * 60)
        logger.info("Starting deployment workflow")
        logger.info("=" * 60)

        # Step 1: Validate prerequisites
        if not self.validate_prerequisites():
            raise Exception("Prerequisites validation failed")

        # Step 2: Upload schemas if bucket specified
        if bucket_name:
            self.upload_schemas(bucket_name)

        # Step 3: Deploy agent
        deployment_info = self.deploy_agent()

        # Step 4: Test deployment
        self.test_deployment(deployment_info['endpoint_url'])

        logger.info("=" * 60)
        logger.info("🎉 Deployment Complete!")
        logger.info("=" * 60)
        logger.info(f"Project: {deployment_info['project_id']}")
        logger.info(f"Region: {deployment_info['region']}")
        logger.info(f"Agent: {deployment_info['agent_name']}")
        logger.info(f"Endpoint: {deployment_info['endpoint_url']}")
        logger.info("=" * 60)

        return deployment_info


def main():
    """Main deployment function"""
    parser = argparse.ArgumentParser(
        description="Deploy document extraction engine to Vertex AI Agent Engine"
    )
    parser.add_argument(
        "--project-id",
        required=True,
        help="Google Cloud project ID"
    )
    parser.add_argument(
        "--region",
        default="us-central1",
        help="GCP region (default: us-central1)"
    )
    parser.add_argument(
        "--config",
        default="config/agent_config.yaml",
        help="Path to agent configuration file"
    )
    parser.add_argument(
        "--bucket",
        help="GCS bucket name for schemas (optional)"
    )

    args = parser.parse_args()

    try:
        deployer = AgentEngineDeployer(
            project_id=args.project_id,
            region=args.region,
            config_path=args.config
        )

        deployment_info = deployer.deploy(bucket_name=args.bucket)

        # Save deployment info
        output_file = "deployment/deployment_info.yaml"
        with open(output_file, 'w') as f:
            yaml.dump(deployment_info, f)

        logger.info(f"Deployment info saved to: {output_file}")

        return 0

    except Exception as e:
        logger.error(f"Deployment failed: {e}")
        import traceback
        traceback.print_exc()
        return 1


if __name__ == "__main__":
    sys.exit(main())
